RestAPI/requirements.txt
anyio==3.3.4
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==8.0.3
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-admin==2.0.1
django-allauth==0.44.0
django-cors-headers==3.11.0
django-crispy-forms==1.11.2
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
future==0.18.2
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httptools==0.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.1.1
orjson==3.6.4
pandas==1.2.4
Pillow==8.2.0
psutil==5.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pystache==0.5.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.19.1
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.3
screen==1.0.1
seaborn==0.11.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
starlette==0.16.0
static==1.1.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
ujson==4.2.0
urllib3==1.26.5
uvicorn==0.15.0
uvloop==0.16.0
waitress==2.0.0
watchgod==0.7
webencodings==0.5.1
websockets==10.0
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlwt==1.3.0
zope.interface==5.4.0
